+++ app/javascript/vue/tasks/otu/browse/const/componentNames.js
@@ -39,6 +39,12 @@ export default {
     otu: true,
     rankGroup: ['SpeciesGroup', 'SpeciesAndInfraspeciesGroup']
   },
+  FieldOccurrences: {
+    title: 'Field occurrences',
+    status: 'prototype',
+    otu: true,
+    rankGroup: ['SpeciesGroup', 'SpeciesAndInfraspeciesGroup']
+  },
   DescriptionComponent: {
     title: 'Description',
     status: 'prototype'