+++ app/javascript/vue/tasks/otu/browse/store/actions/actions.js
@@ -7,6 +7,7 @@ import loadCommonNames from './loadCommonNames'
 import loadConveyances from './loadConveyances'
 import loadDepictions from './loadDepictions'
 import loadDistribution from './loadDistribution'
+import loadFieldOccurrences from './loadFieldOccurrences'
 import loadPreferences from './loadPreferences'
 import loadAssertedDistributions from './loadAssertedDistributions'
 import loadDescendants from './loadDescendants'
@@ -26,6 +27,7 @@ const ActionFunctions = {
   [ActionNames.LoadDescendants]: loadDescendants,
   [ActionNames.LoadDistribution]: loadDistribution,
   [ActionNames.LoadInformation]: loadInformation,
+  [ActionNames.LoadFieldOccurrences]: loadFieldOccurrences,
   [ActionNames.LoadObservationDepictions]: loadObservationDepictions,
   [ActionNames.LoadOtus]: loadOtus,
   [ActionNames.LoadPreferences]: loadPreferences,