+++ app/javascript/vue/tasks/otu/browse/store/getters/getters.js
@@ -11,6 +11,7 @@ import getPreferences from './getPreferences'
 import getUserId from './getUserId'
 import getAssertedDistributions from './getAssertedDistributions'
 import getCurrentOtu from './getCurrentOtu'
+import getFieldOccurrences from './getFieldOccurrences'
 import getLoadState from './getLoadState'
 import getTaxonName from './getTaxonName'
 import getTaxonNames from './getTaxonNames'
@@ -29,6 +30,7 @@ const GetterNames = {
   GetDepictions: 'getDepictions',
   GetDescendants: 'getDescendants',
   GetGeoreferences: 'getGeoreferences',
+  GetFieldOccurrences: 'getFieldOccurrences',
   GetPreferences: 'getPreferences',
   GetUserId: 'getUserId',
   GetAssertedDistributions: 'getAssertedDistributions',
@@ -63,6 +65,7 @@ const GetterFunctions = {
   [GetterNames.GetObservationsDepictions]: getObservationsDepictions,
   [GetterNames.GetRelatedBiologicalAssociations]:
     getRelatedBiologicalAssociations,
+  [GetterNames.GetFieldOccurrences]: getFieldOccurrences,
   [GetterNames.IsSpeciesGroup]: isSpeciesGroup
 }