+++ app/javascript/vue/tasks/otu/browse/store/mutations/setFieldOccurrences.js
@@ -1,3 +1,3 @@
 export default (state, value) => {
-  state.taxon_determinations = value
+  state.fieldOccurrences = value
 }